requirements.txt
flask==2.0.1
werkzeug==2.0.3
numpy==1.24.3
decorator>=4.0.2,<5.0
imageio==2.31.1
imageio-ffmpeg==0.4.8
moviepy==1.0.3
edge-tts==6.1.9
websocket-client==1.6.4
requests>=2.28.0
proglog==0.1.10
tqdm==4.66.1
pyyaml>=6.0
python-dotenv>=0.19.0
flask-cors==3.0.10
redis>=4.5.0 